refactor(lyrics): preserve timestamp-only lines in synced editor - Updated parse_lrc() to keep [mm:ss.xx] lines even when the lyric text is empty (no longer skips timestamp-only rows). - Adjusted parsing to only ignore truly blank lines without timestamps (unchanged behavior for non-synced empties). feat(lyrics): save and render blank synced lines - Updated _emit_save() to include empty-text rows when building the LRC output. - LRC serialization now emits timestamp-only lines as "[mm:ss.xx]" (no trailing space/text). - Plain-text serialization preserves blank lines derived from the synced table.
